// src/app/entities/grade/topic/topic-enum.ts
export enum Topic {
  THEME_OBEYED = 'THEME_OBEYED',
  PROPER_TITLE = 'PROPER_TITLE',
  CONTAINS_INTRODUCTION = 'CONTAINS_INTRODUCTION',
  CONTAINS_DEVELOPMENT = 'CONTAINS_DEVELOPMENT',
  CONTAINS_CONCLUSION = 'CONTAINS_CONCLUSION',
  NUMBER_OF_LINES = 'NUMBER_OF_LINES',
  COHESION = 'COHESION',
  ORTHOGRAPHY = 'ORTHOGRAPHY',
  PUNCTUATION = 'PUNCTUATION'
}